pyproject.toml
[project]
name = "XSS_detector"
version = "0.1.0"
description = "XSS detection Project"
authors = [
    {name = "Your Name", email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "requests>=2.31.0",
    "zss==1.2.0",
    "bs4>=0.0.2",
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.2",
    "pydantic-settings>=2.0.3",
    "jinja2>=3.1.2",
    "markupsafe>=2.1.3",
    "torch==2.3.1",
    "torchvision==0.18.1",
    "torchaudio==2.3.1",
    "matplotlib",
    "pandas",
    "numpy",
    "scikit-learn",
    "gensim (>=4.3.3,<5.0.0)",
    "openai (>=1.69.0,<2.0.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
notebook = "^7.3.3"
ipykernel = "^6.29.5"
cython = "^3.0.12"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
